Release service and callback handler locks before invoking callbacks
Previously, the _ServiceManager.on_request() and
_CallbackManager.fire_message() calls would lock access to the service
and callback objects while callbacks were being delivered. If any
calls from those callbacks were made to add/remove a service or
callback, a thread deadlock could occur.

In this commit, the locks on the service and callback variables from
within the on_request() and fire_message() methods are only held long
enough to get a reference to a copy of the related data which will not
change during delivery of a callback. The locks are released before the
callbacks are invoked, preventing the deadlock.