Avoid use of lock in service and callback handlers
In the previous commit, a lock was obtained around getting the service
and callback objects in _ServiceManager.on_request() and
_CallbackManager.fire_message(), respectively, in order to avoid
possible issues with concurrent iteration and modification activities.

In this commit, modifications to the service and callback objects are
only done to local variables within the methods. Only the service and
callback object references themselves are updated after changes made to
the nested data structures have been completed. With the assumption that
individual variable assignments / reads are done atomically in various
Python interpreters, the locks around the read of the immediate object
references in _ServiceManager.on_request() and
_CallbackManager.fire_message() were removed. This may result in
improved performance.