Wrap self.services.keys call with list in _ServiceManager.destroy
Previously, the Service_Manager.destroy() method used a call to
self.services.keys() to walk over and delete entries from the
dictionary. In Python 3, the .keys() call returns a view into the
dictionary keys rather than a list copy. Because of this, an error would
occur in this code because the size of the dictionary decreases during
iteration.

This commit wraps the .keys() call with a call to list() in order to
create a separate copy of the keys before iteration. This should allow
items to be deleted from the dictionary without encountering the error.