Add setup.cfg and test extras_require to setup.py
This commit adds a setup.cfg with nosetests parameters. This allows
tests to be run via "python setup.py test" with no additional options or
configuration.

This commit also adds a test extras_require which reuses the test
requirements that the setup.py file already uses. This allows for the
test dependencies to be installed from pip via running 'pip install
.[test]".