Avoid exception during remove call made for unknown callback
Previously, if a remove call was made for a callback that was not
currently in the list of registered callbacks for the associated topic
but the associated topic had a different callback registered for it, a
'list.remove(x): x not in list' error occurred. With this commit, the
remove call would return to the caller without raising an error.