Use service handler lock during register/unregister and test fixes
This commit adds a lock around the calls in the
ServiceRegistrationHandler which send register and unregister payloads
to the broker. Previously, without this lock, overlapping
register/unregister calls could lead to unregister calls failing to
perform an unregister - or to perform an unregister but then have the
service implicitly re-registered when a TTL timer callback fires. With
the addition of the lock, these overlapping requests would be
processed serially.

If the ServiceRegistrationHandler unregister call is made before the
service is first registered, a successful unregister response is now
returned to the caller rather than an exception being thrown.

This commit also contains a number of changes to the tests in the
register_service_tests module. The changes include removal of some race
conditions which led to intermittent failures.