Bypass waiting for the thread pool to shutdown from the client destructor
Previously, the shutdown logic for the thread pool would join on the
completion of every thread. If the thread pool were shutdown from the
client destructor during a process shutdown, however, the process could
hang indefinitely (never shutting down).

In this commit, the shutdown logic for the thread pool avoids the thread
join when called from the client destructor, avoiding the hang. Joins on
the thread pool threads, however, will still be done when the client's
destroy() method is called explicitly or if a client instance falls
out of scope after a with() block. Hangs on the thread joins have not
been observed in the latter cases.