Add disconnect wait loop and move reset_current_broker call up
Previously, two disconnect calls made back to back could lead to the
second call trying to perform connection cleanup which was already
partially in process from the first call. Depending upon timing, this
could lead to the second call blocking / hanging indefinitely. In this
commit, a wait loop similar to the one used in the connect() method for
the connection state to change was added to the disconnect() method.
This should significantly reduce the likelihood of redundant/conflicting
disconnect() operations overlapping one another.

Previously, the reset_current_broker() call made to clear out the
current broker was made during processing of the _on_disconnect_run()
call after notification was delivered to condition waiters that a
disconnect had occurred. If immediately acted upon, this could lead to
the consumer seeing a stale value for the current broker. This commit
moves the reset call before the notification is made in order to reduce
latency for the current broker value change.