Make subscribe and unsubscribe synchronous
Previously, the client subscribe and unsubscribe calls were
asynchronous. In this commit, these calls wait for a corresponding ack
to be delivered from the broker before returning control to the caller.
A maximum timeout of 120 seconds is used to wait for an ack. If the
timeout is exceeded, a WaitTimeoutException is raised.