Broker.parse() fixes and tests
Previously, the unique id for a broker constructed via Broker.parse()
was set to None. With this commit, a unique id would be generated for
the broker.

Previously, if a hostname string passed into Broker.parse() contained
both an IPv6 address and a port - e.g., [ff02:1]:8883 - the broker would
fail to be constructed properly. With this commit, the broker would be
constructed with the proper hostname and port.

This commit also includes additional tests for the Broker.parse()
method.