Pre-initialize variables in client constructor
Previously, if an error was raised during the client constructor -- for
example, if a bad configuration parameter or file could not be found
from the configuration parameter -- the _destroy() logic run from the
client destructor would raise a second error for any non-existent
variables in the client object. This commit pre-initializes objects that
may be referenced from the destructor before the logic in the
constructor which might raise an error is run. Logic was also added to
the _destroy() method to not make calls on any objects set to "None".
This should enable the _destroy() logic to not raise another error in
the event that an initial error was encountered during client
construction.