Output more friendly error under Python 3 when CLI subcommand missing
Previously, if one of the optional arguments was supplied for the CLI
(-s or -v) but the subcommand name itself was not, a 'Namespace' does
not have a 'func' error was printed when running under Python 3. This
appears to be due to a change made in the argparser library - see
https://bugs.python.org/issue9253#msg186387. With the changes in this
commit, the presence of the subcommand is validated under both Python 2
and Python 3 even if other optional arguments are supplied.