Fix for timing-specific subscription ack deadlock
Due to changes made in paho-mqtt version 1.4, the DXL client could
encounter a hang during a subscribe or unsubscribe call. The hang would
occur for the following sequence of events:

* Subscribe / unsubscribe call made on the DXL client.
* _packets_awaiting_ack_condition taken in DXL client code.
* Sub/unsub packet sent to the DXL broker.
* While still processing the subscribe / unsubscribe call within the
  Paho MQTT client, the SUB/UNSUB ack packet is received and the
  _callback_mutex is taken.
* On the thread processing the original subscribe / unsubscribe call, an
  attempt is made to take the _callback_mutex. This blocks indefinitely.
* on_sub / on_unsub callback for the SUB/UNSUB ack packet in DXL client
  code tries to take the _packets_awaiting_ack_condition.
* The two threads waiting on the _callback_mutex and
  _packets_awaiting_ack_condition are deadlocked, resulting in the hang.

In this commit, the DXL client code no longer takes the mutex for the
_packets_awaiting_ack_condition before the subscribe / unsubscribe call
is made. Instead, the id for each sub/unsub ack packet is stored in a
set. The subscribe / unsubscribe calls block until the ACK packet id
is inserted into the set. By avoiding conflicting usages of the
_packets_awaiting_ack_condition from different threads, the deadlock
should no longer occur.

This commit also allows paho-mqtt version 1.4 or later to be installed
with the dxlclient again, removing the prior pin to version 1.3 only.