Changed other fields test to ignore extraneous keys
Previously, the message other fields test asserted that the other fields
received for an event message were identical to the other fields in the
original message. In newer DXL brokers, however, some extra keys -- in
particular, a 'dxl.certs' key -- may be added to the other fields
object. This change caused the test to start failing.

In this commit, the message other fields test only makes value
assertions for keys which were included in the original event message.
Any extra keys, such as a 'dxl.certs' key, would now be ignored. This
allows the test to pass again.